Correcting english errors and warnings.

+++ src/Printer.cpp
@@ -11,28 +11,28 @@ Printer::Printer() : lcd(I2C_ADDR, 20, 4)
 
 void Printer::WriteL1(String str, int pos)
 {
-    lcd.setCursor(pos, 0);
+    lcd.setCursor((uint8_t)pos, 0);
     lcd.print(str);
     
 }
 
 void Printer::WriteL2(String str, int pos)
 {
-    lcd.setCursor(pos, 1);
+    lcd.setCursor((uint8_t)pos, 1);
     lcd.print(str);
   
 }
 
 void Printer::WriteL1(int i, int pos)
 {
-    lcd.setCursor(pos, 0);
+    lcd.setCursor((uint8_t)pos, 0);
     lcd.print(i);
 
 }
 
 void Printer::WriteL2(int i, int pos)
 {
-    lcd.setCursor(pos, 1);
+    lcd.setCursor((uint8_t)pos, 1);
     lcd.print(i);
 
 }
